restructure EmbeddedAssets functionality into its own class with more distinct subroutines
